checkpoint callback

+++ ml/load_n_train.py
@@ -42,43 +42,45 @@ model_path = config['model_weights']
 
 bad_users = json.loads(os.getenv("BAD_USERS"))
 
-folders = []
-folders.append(img_folder)
-for dir in folders:
-    if os.path.exists(dir):
-        shutil.rmtree(dir)
-    os.makedirs(dir)
 
-'''
-Initializes the classmap of concept names to training id's.
-(these id's don't represent the conceptid's from our database)
-'''
-start = time.time()
-print("Initializing Classmap.")
+if True: 
+    folders = []
+    folders.append(img_folder)
+    for dir in folders:
+        if os.path.exists(dir):
+            shutil.rmtree(dir)
+        os.makedirs(dir)
 
-classmap = []
-for concept in concepts:
-    name = queryDB("select name from concepts where id=" + str(concept)).iloc[0]["name"]
-    classmap.append([name,concepts.index(concept)])
-classmap = pd.DataFrame(classmap)
-classmap.to_csv(class_map_file,index=False, header=False)
-classmap = classmap.to_dict()[0]
+    '''
+    Initializes the classmap of concept names to training id's.
+    (these id's don't represent the conceptid's from our database)
+    '''
+    start = time.time()
+    print("Initializing Classmap.")
 
-end = time.time()
-print("Done Initializing Classmap: " + str((end - start)/60) + " minutes")
+    classmap = []
+    for concept in concepts:
+        name = queryDB("select name from concepts where id=" + str(concept)).iloc[0]["name"]
+        classmap.append([name,concepts.index(concept)])
+    classmap = pd.DataFrame(classmap)
+    classmap.to_csv(class_map_file,index=False, header=False)
+    classmap = classmap.to_dict()[0]
+
+    end = time.time()
+    print("Done Initializing Classmap: " + str((end - start)/60) + " minutes")
 
 
-'''
-Downloads the annotation data and saves it into training and validation csv's.
-Also downloads corresponding images.
-'''
-start = time.time()
-print("Starting Download.")
+    '''
+    Downloads the annotation data and saves it into training and validation csv's.
+    Also downloads corresponding images.
+    '''
+    start = time.time()
+    print("Starting Download.")
 
-download_annotations(min_examples, concepts, classmap, bad_users, img_folder, train_annot_file, valid_annot_file)
+    download_annotations(min_examples, concepts, classmap, bad_users, img_folder, train_annot_file, valid_annot_file)
 
-end = time.time()
-print("Done Downloading Annotations: " + str((end - start)/60) + " minutes")
+    end = time.time()
+    print("Done Downloading Annotations: " + str((end - start)/60) + " minutes")
 
 
 '''
@@ -88,7 +90,7 @@ start = time.time()
 print("Starting Training.")
 
 model = models.backbone('resnet50').retinanet(num_classes=len(concepts), modifier=freeze_model)
-model = multi_gpu_model(model, gpu=4)
+# model = multi_gpu_model(model, gpu=4)
 model.load_weights(model_path, by_name=True, skip_mismatch=True)
 
 model.compile(
@@ -119,9 +121,19 @@ test_generator = CSVGenerator(
     valid_annot_file,
     class_map_file,
 )
-model.fit_generator(train_generator, epochs=10, validation_data=test_generator)
+
+# Checkpoint: save models that are improvements
+filepath = "weights/weights-{epoch:02d}-{val_acc:.4f}.h5"
+checkpoint = ModelCheckpoint(filepath, monitor='val_acc', verbose=0, save_best_only=True, mode='max')
+
+history = model.fit_generator(train_generator, 
+    epochs=10, 
+    callbacks=[checkpoint],
+    validation_data=test_generator
+    ).history
 
 end = time.time()
 print("Done Training Model: " + str((end - start)/60) + " minutes")
+print(history)
